refactor: move enoent to its own module
